pyproject.toml
[project]
dependencies = [
    "beautifulsoup4",
    "celery",
    "django~=4.2",
    "django-allauth",
    "django-braces",
    "django-celery-beat",
    "django-debug-toolbar",
    "django-environ",
    "django-extensions",
    "django-filter==25.1",
    "django-handyhelpers",
    "flower",
    "PyJWT<3",
    "psycopg2-binary",
    "redis",
    "requests>=2.26.0",
    "pillow",
    "playwright",
    "whitenoise",
    ]
description = "Inland Northwest Tech Community"
dynamic = ["version"]
keywords = ["django"]
license = {file = "LICENSE"}
name = "SpokaneTech"
readme = "README.md"
requires-python = ">=3.8"


[project.optional-dependencies]
dev = [
    "bandit",
    "black",
    "coveralls",
    "django-debug-toolbar",
    "faker",
    "isort",
    "model-bakery",
    "mypy",
    "mypy-extensions",
    "ruff",
    "pytest",
    "pytest-cov",
    "pytest-django",
    "radon",
    "safety",
    "types-python-dateutil",
    "types-requests",
    "typing_extensions",
    ]
docker = [
    "gunicorn",
]

[tool.bandit]
exclude_dirs = [
    "venv", 
    "src/django_project/local_test", 
    "src/django_project/tests", 
    "src/django_project/*/management/commands",
    "src/django_project/*/scripts/*"
    ]


[tool.black]
line-length = 120


[tool.coverage.report]
show_missing = true
omit = [
    "src/django_project/manage.py", 
    "src/django_project/core/asgi.py",
    "src/django_project/core/wsgi.py", 
    "src/django_project/*/migrations/*",
    "src/django_project/*/scripts/*",
    "src/django_project/tests/*", 
    ]


[tool.coverage.run]
branch = true
source = ["src/django_project"]


[tool.isort]
profile = "black"


[tool.mypy]
exclude = [
    "venv/*", 
    "src/django_project/local_test",
    "src/django_project/.*/migrations/.*",
    "src/django_project/.*/scripts/.*",
    "src/django_project/tests/.*",
    ]
ignore_missing_imports = true


[tool.pytest.ini_options]
addopts = "-s -v -x --strict-markers -m 'not extra'"
testpaths = ["src/django_project"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::UserWarning",
    ]


[tool.ruff]
line-length = 120
exclude = [
    "src/django_project/manage.py", 
    "src/django_project/tests", 
    "src/django_project/*/migrations", 
    "src/django_project/*/scripts", 
    "src/django_project/*/local_test"
    ]


[tool.setuptools.packages.find]
where = ["."]
include = ["src/django_project"]
exclude = []
namespaces = false
